# Models/Requests/UIComponentGeneratorRequest.cs
namespace Utils.IO.Server.Models.Requests
{
    public class UIComponentGeneratorRequest
    {
        public string? NameOfComponent { get; set; }
        public string? WebFrameworkUsed { get; set; }
        public string? UILibraryUsed { get; set; }
        public string? PurposeOfComponent { get; set; }
        public string? HowIsComponentIsStyled { get; set; }
        public string? Web3WalletFeatureUsed { get; set; }
    }
}