# swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Utils IO Server",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/Gpt": {
      "get": {
        "tags": [
          "Gpt"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Gpt/article-summarizer": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/paragraph-generator": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommonGptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/convert-code": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCodeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/smart-contract-generator": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartContractGeneratorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartContractGeneratorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SmartContractGeneratorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/ui-component-generator": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UIComponentGeneratorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UIComponentGeneratorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UIComponentGeneratorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/text-to-image-generator": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextToImageGeneratorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TextToImageGeneratorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TextToImageGeneratorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Gpt/chatbot": {
      "post": {
        "tags": [
          "Gpt"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Polling/status/{gptId}": {
      "get": {
        "tags": [
          "Polling"
        ],
        "parameters": [
          {
            "name": "gptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Polling/results/{gptId}": {
      "get": {
        "tags": [
          "Polling"
        ],
        "parameters": [
          {
            "name": "gptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Stripe": {
      "get": {
        "tags": [
          "Stripe"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Stripe/webhook": {
      "post": {
        "tags": [
          "Stripe"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatbotMessage": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChatbotRequest": {
        "type": "object",
        "properties": {
          "newMessage": {
            "type": "string",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatbotMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CommonGptRequest": {
        "type": "object",
        "properties": {
          "inputText": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConvertCodeRequest": {
        "type": "object",
        "properties": {
          "codeToConvert": {
            "type": "string",
            "nullable": true
          },
          "fromProgrammingLanguage": {
            "type": "string",
            "nullable": true
          },
          "toProgrammingLanguage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResultsResponse": {
        "type": "object",
        "properties": {
          "resultUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SmartContractGeneratorRequest": {
        "type": "object",
        "properties": {
          "blockchain": {
            "type": "string",
            "nullable": true
          },
          "tokenStandard": {
            "type": "string",
            "nullable": true
          },
          "contractName": {
            "type": "string",
            "nullable": true
          },
          "whatDoesItDo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TextToImageGeneratorRequest": {
        "type": "object",
        "properties": {
          "concept": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UIComponentGeneratorRequest": {
        "type": "object",
        "properties": {
          "nameOfComponent": {
            "type": "string",
            "nullable": true
          },
          "webFrameworkUsed": {
            "type": "string",
            "nullable": true
          },
          "uiLibraryUsed": {
            "type": "string",
            "nullable": true
          },
          "purposeOfComponent": {
            "type": "string",
            "nullable": true
          },
          "howIsComponentIsStyled": {
            "type": "string",
            "nullable": true
          },
          "web3WalletFeatureUsed": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkResponse": {
        "type": "object",
        "properties": {
          "workGuid": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}